Finalizado teste de componente e unitário para delecao de clientes

+++ src/main/java/com/brasilprev/customermanagement/delete/usecase/DeleteClientUseCase.java
@@ -3,6 +3,7 @@ package com.brasilprev.customermanagement.delete.usecase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.brasilprev.customermanagement.commons.usecase.exception.DomainNotFoundException;
 import com.brasilprev.customermanagement.delete.usecase.gateway.DeleteClientGateway;
 
 @Service
@@ -12,6 +13,9 @@ public class DeleteClientUseCase {
 	private DeleteClientGateway deleteClientGateway;
 
 	public void deleteClient(Long clientId) {
+		if(!deleteClientGateway.existClientById(clientId))
+			throw new DomainNotFoundException();
+		
 		deleteClientGateway.deleteClientById(clientId);
 	}
